# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Bello app auth",
    "description": "An API for bello app authentication"
  },
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
...
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/v1/user/login": {
      "post": {
        "description": "Login user",
        "tags": [
          "User"
        ],
        "operationId": "loginUser",
        "parameters": [
          {
            "name": "Login",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully registered",
            "schema": {
              "$ref": "#/definitions/TokenResponse"
            }
          },
          "400": {
            "description": "Bad request payload",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Service down",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "message": {
          "type": "string",
          "enum": [
            "InternalServerError",
            "Unauthorized",
            "NotFound",
            "UserAlreadyExists",
            "IncorrectCredentials",
            "BadRequest"
          ]
        }
      }
    },
    "TokenResponse": {
      "properties": {
        "accessToken": {
          "type": "string",
          "x-nullable": false
        },
        "refreshToken": {
          "type": "string",
          "x-nullable": false
        }
      }
    },
...
    "LoginRequest": {
      "type": "object",
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string",
          "x-nullable": false
        },
        "password": {
          "type": "string",
          "x-nullable": false
        }
      }
    },
...
  }
}